separating submodules in backend, added env variables for more secure serving
