+++ backend/go.mod
@@ -3,15 +3,14 @@ module word-roulette_go
 go 1.21.1
 
 require (
-	github.com/google/uuid v1.4.0
-	github.com/gorilla/websocket v1.5.0
-	github.com/redis/go-redis/v9 v9.3.0
+	github.com/google/uuid v1.6.0
+	github.com/gorilla/websocket v1.5.1
+	github.com/redis/go-redis/v9 v9.4.0
 	github.com/rs/cors v1.10.1
 )
 
 require (
 	github.com/cespare/xxhash/v2 v2.2.0 // indirect
 	github.com/dgryski/go-rendezvous v0.0.0-20200823014737-9f7001d12a5f // indirect
-	github.com/felixge/httpsnoop v1.0.1 // indirect
-	github.com/gorilla/handlers v1.5.1 // indirect
+	golang.org/x/net v0.21.0 // indirect
 )
