+++ frontend/package.json
@@ -11,7 +11,6 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "dotenv": "^16.4.1",
     "localtunnel": "^2.0.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -24,6 +23,7 @@
     "@babel/preset-react": "^7.22.15",
     "babel-loader": "^9.1.3",
     "css-loader": "^6.8.1",
+    "dotenv-webpack": "^8.0.1",
     "html-webpack-plugin": "^5.5.3",
     "sass-loader": "^13.3.2",
     "style-loader": "^3.3.3",
