+++ frontend/src/components/App.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useRef, useState } from 'react';
-import Lobby from './Lobby'
+import Lobby from './Lobby';
 
 const colorList = [
   'rgb(255, 87, 51)',    // Reddish-Orange
@@ -21,10 +21,9 @@ const App = () => {
   const socket = useRef(null);
   
   useEffect(() => {
-    // need to determine external vs localhost
-    // external --> Create .env file in the frontend directory and assign 'EXT_IP' the value of the external IP of the machine the go server is running on
-    // TEMPORARY SOLUTION IS TO HARD CODE EXTERNAL IP in place of `your_external_ip`
-    socket.current = new WebSocket(`ws://your_external_ip/ws`) //|| new WebSocket('ws://localhost:8085/ws');
+    // assign external IP to `EXT_IP` in a .env created in frontend dir
+    // add the .env to your `.gitignore` to avoid pushing it to GitHub
+    socket.current = new WebSocket(`ws://${process.env.EXT_IP}/ws`) //|| new WebSocket('ws://localhost:8085/ws');
 
     socket.current.addEventListener('open', (e) => {
       console.log('WebSocket connected');
