+++ frontend/webpack.config.js
@@ -1,5 +1,6 @@
 const path = require('path');
 const HTMLWebpackPlugin = require('html-webpack-plugin');
+const dotenv = require('dotenv-webpack');
 
 module.exports = {
   // was needed when trying dotenv, unsure of reason yet
@@ -21,6 +22,7 @@ module.exports = {
       title: 'Development',
       template: './src/index.html', // Path to HTML template
     }),
+    new dotenv() // make .env variables available to webpack when bundling
   ],
 
   devServer: {
